+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## [version 1.6.3] 2024-03-31
 
+### Changed
+
+- Detail panel is now displaying the ReadOnly object fields
+
+### Updated
+
 - Update package [file_picker: 8.0.0]
 - Update package [fl_chart: ^0.67.0]